♻️ 96X7WA code: deduplicate built-in blueprint builders
