+++ .agentplane/tasks/202605091753-96X7WA/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605091753-96X7WA/blueprint-builder",
+  "created_at": "2026-05-09T17:55:52.997Z",
+  "head_sha": "e2218d7ad7b1d98bcb75d40ba769a5cf8cb755ef",
+  "last_verified_at": "2026-05-09T18:07:21.515Z",
+  "last_verified_sha": "e2218d7ad7b1d98bcb75d40ba769a5cf8cb755ef",
+  "schema_version": 1,
+  "task_id": "202605091753-96X7WA",
+  "updated_at": "2026-05-09T17:55:52.997Z",
+  "verify": {
+    "status": "pass"
+  }
+}
