+++ .agentplane/tasks/202605091753-96X7WA/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-09T17:55:52.997Z
+
+## Task
+
+- Task: `202605091753-96X7WA`
+- Title: Deduplicate built-in blueprint builders
+- Status: DOING
+- Branch: `task/202605091753-96X7WA/blueprint-builder`
+- Canonical task record: `.agentplane/tasks/202605091753-96X7WA/README.md`
+
+## Verification
+
+- State: ok
+- Note: Verified: extracted shared blueprint builder and branch_pr route helpers; focused blueprint tests passed (3 files, 51 tests), typecheck passed, clone:report improved duplicatedLines 1708->1546 and duplicatedTokens 17574->16193, and clone:check passed without baseline update.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-09T17:55:52.997Z
+- Branch: task/202605091753-96X7WA/blueprint-builder
+- Head: e2218d7ad7b1
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
