+++ packages/agentplane/src/blueprints/builtins.ts
@@ -1,147 +1,7 @@
-import type {
-  Blueprint,
-  BlueprintContextBudget,
-  BlueprintEdge,
-  BlueprintId,
-  BlueprintNode,
-  BlueprintNodeKind,
-  EvidenceKind,
-  EvidenceRequirement,
-  StopRule,
-  TaskKind,
-  WorkflowMode,
-} from "./model.js";
+import type { Blueprint } from "./model.js";
 import { SPECIALIZED_CODE_BLUEPRINTS } from "./builtins-specialized.js";
-
-type NodeSpec = {
-  kind: BlueprintNodeKind;
-  mode?: BlueprintNode["mode"];
-  evidence?: readonly EvidenceKind[];
-  protected?: boolean;
-  allowedCommands?: readonly string[];
-  policyModules?: readonly string[];
-};
-
-const CORE_STOP_RULES: readonly StopRule[] = [
-  {
-    id: "required_evidence_missing",
-    severity: "stop",
-    reason: "Required blueprint evidence cannot be produced.",
-  },
-  {
-    id: "protected_lifecycle_override",
-    severity: "stop",
-    reason: "A route or recipe attempts to override protected lifecycle behavior.",
-  },
-];
-
-const NODE_MODE_BY_KIND = {
-  intake: "deterministic",
-  scope: "deterministic",
-  context_resolve: "deterministic",
-  approval_gate: "approval",
-  worktree_start: "deterministic",
-  work_unit: "agentic",
-  deterministic_check: "deterministic",
-  fast_local_checks: "deterministic",
-  artifact_write: "deterministic",
-  pr_artifact: "deterministic",
-  hosted_checks: "deterministic",
-  publish_or_integrate: "deterministic",
-  verify_record: "record",
-  handoff: "record",
-  finish: "deterministic",
-} as const satisfies Record<BlueprintNodeKind, BlueprintNode["mode"]>;
-
-function node(spec: NodeSpec): BlueprintNode {
-  return {
-    id: spec.kind,
-    kind: spec.kind,
-    mode: spec.mode ?? NODE_MODE_BY_KIND[spec.kind],
-    required: true,
-    ...(spec.protected ? { protected: true } : {}),
-    ...(spec.allowedCommands ? { allowedCommands: spec.allowedCommands } : {}),
-    ...(spec.policyModules ? { policyModules: spec.policyModules } : {}),
-    ...(spec.evidence ? { evidence: spec.evidence } : {}),
-  };
-}
-
-function edgesFor(nodes: readonly BlueprintNode[]): BlueprintEdge[] {
-  return nodes.slice(0, -1).map((current, index) => {
-    const next = nodes[index + 1];
-    if (!next) {
-      throw new Error(`Blueprint edge construction failed after ${current.id}`);
-    }
-    return { from: current.id, to: next.id };
-  });
-}
-
-function evidence(
-  id: string,
-  kind: EvidenceKind,
-  producedBy: BlueprintNodeKind,
-  description: string,
-): EvidenceRequirement {
-  return {
-    id,
-    kind,
-    producedBy,
-    required: true,
-    description,
-  };
-}
-
-function blueprint(opts: {
-  id: BlueprintId;
-  title: string;
-  description: string;
-  taskKinds: readonly TaskKind[];
-  workflowModes?: readonly WorkflowMode[];
-  allowedCommands: readonly string[];
-  policyModules: readonly string[];
-  contextBudget: BlueprintContextBudget;
-  nodes: readonly BlueprintNode[];
-  requiredEvidence: readonly EvidenceRequirement[];
-  stopRules?: readonly StopRule[];
-}): Blueprint {
-  return {
-    id: opts.id,
-    version: 1,
-    title: opts.title,
-    description: opts.description,
-    taskKinds: opts.taskKinds,
-    ...(opts.workflowModes ? { workflowModes: opts.workflowModes } : {}),
-    allowedCommands: opts.allowedCommands,
-    policyModules: opts.policyModules,
-    contextBudget: opts.contextBudget,
-    nodes: opts.nodes,
-    edges: edgesFor(opts.nodes),
-    requiredEvidence: opts.requiredEvidence,
-    stopRules: [...CORE_STOP_RULES, ...(opts.stopRules ?? [])],
-    recipeExtensionPoints: [
-      {
-        nodeKind: "context_resolve",
-        allowed: ["context_hint", "risk_hint"],
-        rejected: ["policy_override", "permission_bypass"],
-      },
-      {
-        nodeKind: "work_unit",
-        allowed: ["output_schema", "artifact_template"],
-        rejected: ["lifecycle_replacement"],
-      },
-      {
-        nodeKind: "deterministic_check",
-        allowed: ["check_suggestion"],
-        rejected: ["required_check_disable"],
-      },
-      {
-        nodeKind: "verify_record",
-        allowed: ["evidence_requirement"],
-        rejected: ["verification_bypass"],
-      },
-    ],
-  };
-}
+import { blueprint, evidence, node } from "./builtin-builder.js";
+import { codeBranchPrNodes } from "./builtin-routes.js";
 
 const analysisNodes = [
   node({ kind: "intake" }),
@@ -210,48 +70,6 @@ const codeDirectNodes = [
     kind: "deterministic_check",
     evidence: ["check_result"],
     allowedCommands: ["agentplane task verify-show <task-id>", "project focused checks"],
-  }),
-  node({ kind: "verify_record", evidence: ["check_result"], protected: true }),
-  node({ kind: "finish", evidence: ["commit"], protected: true }),
-] as const;
-
-const codeBranchPrNodes = [
-  node({ kind: "intake" }),
-  node({ kind: "scope", evidence: ["assumptions"] }),
-  node({ kind: "approval_gate", evidence: ["approval"], protected: true }),
-  node({
-    kind: "context_resolve",
-    evidence: ["context_manifest"],
-    policyModules: [
-      ".agentplane/policy/security.must.md",
-      ".agentplane/policy/dod.core.md",
-      ".agentplane/policy/dod.code.md",
-      ".agentplane/policy/workflow.branch_pr.md",
-    ],
-  }),
-  node({
-    kind: "worktree_start",
-    evidence: ["changed_paths"],
-    protected: true,
-    allowedCommands: ["agentplane work start <task-id> --agent <ROLE> --slug <slug> --worktree"],
-  }),
-  node({ kind: "work_unit", evidence: ["changed_paths"] }),
-  node({
-    kind: "fast_local_checks",
-    evidence: ["check_result"],
-    allowedCommands: ["agentplane task verify-show <task-id>", "project focused checks"],
-  }),
-  node({
-    kind: "pr_artifact",
-    evidence: ["artifact", "external_link"],
-    allowedCommands: ["agentplane pr open <task-id> --branch <branch> --author <ROLE>"],
-  }),
-  node({ kind: "hosted_checks", evidence: ["check_result", "external_link"] }),
-  node({
-    kind: "publish_or_integrate",
-    evidence: ["commit", "external_link"],
-    protected: true,
-    allowedCommands: ["agentplane integrate <task-id> --branch <branch> --run-verify"],
   }),
   node({ kind: "verify_record", evidence: ["check_result"], protected: true }),
   node({ kind: "finish", evidence: ["commit"], protected: true }),
